requirements/dev.txt
-r base.txt

# Django & Admin Panel
django==5.0.1
djangorestframework==3.14.0
django-filter==23.5
django-cors-headers==4.3.1
gunicorn==21.2.0
aiohttp==3.9.1

# non mysqlclient - just SQLite!
